+++ test/iterators.jl
@@ -1001,3 +1001,42 @@ end
     end
     @test v == ()
 end
+
+@testset "Unfold" begin
+    @test isempty(Iterators.unfold(identity, nothing))
+
+    unfold61 = Iterators.unfold(1) do x
+         if x < 2^26
+            return x, x+x
+         else
+            return nothing
+         end
+    end
+    @test all(enumerate(unfold61)) do (i,x)
+        2^(i-1) == x
+    end
+
+    aa, bb = eachcol(randn(11,2))
+    vals = map(aa, bb) do a, b iterate(Iterators.unfold(a) do x x, b end) end
+    @test all(zip(aa, bb) .== vals)
+
+    struct init_sentinel end
+    @testset "Unfold replicates `iterate` calls" for myitr in [1:2:4, (-5:5)*π, (), (1,2), randn(5)]
+        myUnfold = Iterators.unfold(init_sentinel()) do state
+            if state isa init_sentinel
+                return iterate(myitr)
+            else
+                return iterate(myitr, state)
+            end
+        end
+        @test Iterators.map(==, myUnfold, myitr) |> all
+    end
+
+    fibs = Iterators.unfold(Int64.((1,1))) do (a,b) a, (b, a+b) end
+    fibO1(n) = (MathConstants.φ^n - (1-MathConstants.φ)^n) / √5
+    @test 93 == Iterators.take(Iterators.map(
+        !isapprox,
+        fibs,
+        (fibO1(x) for x in Iterators.countfrom(1))
+    ), 111) |> collect |> findfirst
+end